refactor(settings/ai-providers): assemble redesigned page
Wire the new orchestrator index, swap the route import from
org-ai-providers to ai-providers, delete the 1789-line old view, and
update NoAiProviderEmptyState to use ConnectProviderDialog instead of
the removed ProviderCardGrid.
