+++ apps/mesh/src/web/routes/orgs/settings/ai-providers.tsx
@@ -1,4 +1,4 @@
-import { OrgAiProvidersPage } from "@/web/views/settings/org-ai-providers";
+import { OrgAiProvidersPage } from "@/web/views/settings/ai-providers";
 
 export default function AiProvidersRoute() {
   return <OrgAiProvidersPage />;
